Introduction of the testCasesMode
Depending on the nature of the problem, traditional GP test
cases may or may not be used. The latter happens for example
when there is more than single answer for the given input,
or when problem does not have single invocation property. In
both cases we need to resort to the solver.

This commit adds automatic determination of the mode which
can be used, and 'gp' mode of standard test cases will be
used whenever possible. Mode with solver usage is called
'solver'.

NOTE: this commit focuses on the determination mechanism,
and not every part of engine is yet aware of this distinction.